Ensure resolution results are cached
