+++ rust/saturn/src/lib.rs
@@ -2,6 +2,7 @@ pub mod errors;
 pub mod indexing;
 pub mod model;
 pub mod offset;
+pub mod resolution;
 pub mod source_location;
 pub mod stats;
 pub mod visualization;
